+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/token/generator/impl/OrderByTokenGenerator.java
@@ -27,7 +27,7 @@ import org.apache.shardingsphere.sharding.rewrite.token.generator.IgnoreForSingl
 import org.apache.shardingsphere.sql.parser.sql.segment.generic.table.SimpleTableSegment;
 import org.apache.shardingsphere.sql.parser.sql.statement.dml.SelectStatement;
 import org.apache.shardingsphere.underlying.rewrite.sql.token.generator.OptionalSQLTokenGenerator;
-import org.apache.shardingsphere.sharding.rewrite.token.pojo.impl.OrderByToken;
+import org.apache.shardingsphere.sharding.rewrite.token.pojo.OrderByToken;
 
 /**
  * Order by token generator.
@@ -58,7 +58,7 @@ public final class OrderByTokenGenerator implements OptionalSQLTokenGenerator<Se
         }
         return result;
     }
-
+    
     private int generateOrderByIndex(final SelectStatementContext selectStatementContext) {
         if (selectStatementContext.getGroupByContext().getLastIndex() > 0) {
             return selectStatementContext.getGroupByContext().getLastIndex() + 1;
