+++ sharding-core/sharding-core-rewrite/src/main/java/org/apache/shardingsphere/sharding/rewrite/token/pojo/LiteralGeneratedKeyAssignmentToken.java
@@ -15,19 +15,22 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.sharding.rewrite.token.pojo.impl;
+package org.apache.shardingsphere.sharding.rewrite.token.pojo;
 
 /**
- * Generated key assignment token for parameter marker.
+ * Generated key assignment token for literal.
  */
-public final class ParameterMarkerGeneratedKeyAssignmentToken extends GeneratedKeyAssignmentToken {
+public final class LiteralGeneratedKeyAssignmentToken extends GeneratedKeyAssignmentToken {
     
-    public ParameterMarkerGeneratedKeyAssignmentToken(final int startIndex, final String columnName) {
+    private final Object value;
+    
+    public LiteralGeneratedKeyAssignmentToken(final int startIndex, final String columnName, final Object value) {
         super(startIndex, columnName);
+        this.value = value;
     }
     
     @Override
     protected String getRightValue() {
-        return "?";
+        return value instanceof String ? String.format("'%s'", value) : value.toString();
     }
 }
